// src/app/modules/clients/enums/client.enum.ts
export enum ClientActionTypes {
  FETCH_CLIENT = '[CLIENTS] FETCH_CLIENT',
  FETCH_CLIENT_SUCCESS = '[CLIENTS] FETCH_CLIENT_SUCCESS',
  FETCH_CLIENT_FAILURE = '[CLIENTS] FETCH_CLIENT_FAILURE',
  CREATE_CLIENT = '[CLIENTS] CREATE_CLIENT',
  CREATE_CLIENT_SUCCESS = '[CLIENTS] CREATE_CLIENT_SUCCESS',
  CREATE_CLIENT_FAILURE = '[CLIENTS] CREATE_CLIENT_FAILURE',
  UPDATE_CLIENT = '[CLIENTS] UPDATE_CLIENT',
  UPDATE_CLIENT_SUCCESS = '[CLIENTS] UPDATE_CLIENT_SUCCESS',
  UPDATE_CLIENT_FAILURE = '[CLIENTS] UPDATE_CLIENT_FAILURE',
  DELETE_CLIENT = '[CLIENTS] DELETE_CLIENT',
  DELETE_CLIENT_SUCCESS = '[CLIENTS] DELETE_CLIENT_SUCCESS',
  DELETE_CLIENT_FAILURE = '[CLIENTS] DELETE_CLIENT_FAILURE',
}